extract MagicApiaryInventory into file
